+++ src/Exception/ClientException.php
@@ -2,6 +2,8 @@
 
 namespace Scriptotek\Alma\Exception;
 
-class ClientException extends \RuntimeException
+use RuntimeException;
+
+class ClientException extends RuntimeException
 {
 }