+++ src/Exception/InvalidQueryException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Scriptotek\Alma\Exception;
-
-class InvalidQueryException extends ClientException
-{
-}